Move `choices.py` from utils to core
